[hardpy] Minor refactoring
